create account page and completes auth flow to get user id
